Complete ADS snapshot cache implementation
