+++ pkg/envoy/ads/cache_callbacks.go
@@ -7,6 +7,7 @@ import (
 	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
 
 	"github.com/openservicemesh/osm/pkg/envoy"
+	"github.com/openservicemesh/osm/pkg/messaging"
 	"github.com/openservicemesh/osm/pkg/metricsstore"
 	"github.com/openservicemesh/osm/pkg/utils"
 )
@@ -44,18 +45,59 @@ func (s *Server) OnStreamOpen(ctx context.Context, streamID int64, typ string) e
 	}
 
 	s.proxyRegistry.RegisterProxy(proxy)
+	go func() {
+		// Register for proxy config updates broadcasted by the message broker
+		proxyUpdatePubSub := s.msgBroker.GetProxyUpdatePubSub()
+		proxyUpdateChan := proxyUpdatePubSub.Sub(messaging.ProxyUpdateTopic, messaging.GetPubSubTopicForProxyUUID(proxy.UUID.String()))
+		defer s.msgBroker.Unsub(proxyUpdatePubSub, proxyUpdateChan)
+
+		certRotations, unsubRotations := s.certManager.SubscribeRotations(proxy.Identity.String())
+		defer unsubRotations()
+
+		for {
+			select {
+			case <-proxyUpdateChan:
+				log.Debug().Str("proxy", proxy.String()).Msg("Broadcast update received")
+				s.update(proxy)
+			case <-certRotations:
+				log.Debug().Str("proxy", proxy.String()).Msg("Certificate has been updated for proxy")
+				s.update(proxy)
+			case <-ctx.Done():
+				return
+			}
+		}
+	}()
 	return nil
+}
+
+func (s *Server) update(proxy *envoy.Proxy) {
+	ch := s.workqueues.AddJob(&proxyResponseJob{
+		proxy:     proxy,
+		xdsServer: s,
+		typeURIs:  envoy.XDSResponseOrder,
+		done:      make(chan struct{}),
+	})
+	<-ch
+	close(ch)
 }
 
 // OnStreamClosed is called on stream closed
 func (s *Server) OnStreamClosed(streamID int64) {
 	log.Debug().Msgf("OnStreamClosed id: %d", streamID)
 	s.proxyRegistry.UnregisterProxy(streamID)
+
+	metricsstore.DefaultMetricsStore.ProxyConnectCount.Dec()
 }
 
-// OnStreamRequest is called when a request happens on an open string
-func (s *Server) OnStreamRequest(a int64, req *discovery.DiscoveryRequest) error {
+// OnStreamRequest is called when a request happens on an open connection
+func (s *Server) OnStreamRequest(streamID int64, req *discovery.DiscoveryRequest) error {
 	log.Debug().Msgf("OnStreamRequest node: %s, type: %s, v: %s, nonce: %s, resNames: %s", req.Node.Id, req.TypeUrl, req.VersionInfo, req.ResponseNonce, req.ResourceNames)
+
+	proxy := s.proxyRegistry.GetConnectedProxy(streamID)
+	if proxy != nil {
+		metricsstore.DefaultMetricsStore.ProxyXDSRequestCount.WithLabelValues(proxy.UUID.String(), proxy.Identity.String(), req.TypeUrl).Inc()
+	}
+
 	return nil
 }
